Fix a potential thread dead lock issue
When bigo_worker_thread does kref_put on an instance, if the refcount
changes from 1 to 0, bigo_close will be invoked. If this is the last
instance, on_last_inst_close will be invoked and it will wait for
bigo_worker_thread to exit. It is a dead loop that bigo_worker_thread
is waiting for itself to exit. The fix is to make sure the worker thread
is stopped before closing the last instance.

Bug: 229580907

Signed-off-by: Ruofei Ma <ruofeim@google.com>
Change-Id: I0950a07e072c6d10ab76389bd10ff697c3d1fc79