Set instance to idle at open
By setting instance to idle at open, we a delaying the
power/freq vote until there is activity from this particular
instance. This change can avoid unnecessary power comsumption
on the instances that are opened but have no activity.

Without the change the statement above only works for the
first instance. With this change it works for all instances.
Refer to the bug for details.

Bug: 242975531

Signed-off-by: Ruofei Ma <ruofeim@google.com>
Change-Id: Ia871e9428995fea91a322f9245977210d2da39d9