drm: samsung: limit max_disp_freq for concurrent writeback
When concurrent writeback is enabled, writeback instant off may occur
if the outfifo for writeback is full meanwhile the outfifo for LCD is
not full yet.

This issue could be solved by increasing outfifo size for writeback and
decreasing outfifo size for LCD, but it might cause higher chance to
underrun during normal cases.

So, to avoid it, we can also limit max_disp_freq only when concurrent
writeback is enabled.

Bug: 232887601
Test: VtsHalGraphicsComposer3_TargetTest
Change-Id: I3e663d9cf50f3b31df848fe7546a4bd17235b95f
Signed-off-by: Wiwit Rifa'i <wiwitrifai@google.com>