drm: samsung: protect to init and deinit DSIM/DPHY hardware.
Multiple DSIM and DPHY can be initialized or deinitialized
at the same time. So, it causes problem.

For example, if DSIM0 and DSIM1 are initialized at the same time,
PLL initialization can be failed because of shared SFRs
and DSIM/DPHY hardware initialization sequence.

So, DSIM global mutex lock protects multiple DSIM/DPHY initialization
or deinitialization.

Bug: 257335702
Change-Id: Ib1390ff8b6f20fc2efcb535ba83ac14cbffb17db
Signed-off-by: YongWook Shin <yongwook.shin@samsung.com>
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>