mali_kbase: platform: dvfs: Fix pid struct and task_struct reference leaks
In the process of obtaining the UID for the current process, a reference to
the task's pid struct is leaked via find_get_pid(), in addition to the
task's task_struct. This causes the pid struct and task_struct to remain in
memory forever, as their reference counts will never reach zero, which
constitutes an awful memory leak.

Since kctx->kprcs->tgid is always set to current->tgid, just get the UID
from the current task directly instead of doing a reference-grabbing
lookup. This fixes the reference leaks and subsequent memory leaks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>