gsa: Fix incorrect request length when copying a TZ message request
The length passed to memcpy in gsa_tz_chan_msg_xchg() is blatantly wrong;
`req_len` should be used as-is. This causes out-of-bounds read/write and
can also result in a message request not being copied in its entirety.

Fix it by using `req_len` instead of `sizeof(req_len)` for the memcpy.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>