aoc: Block "logging" service since it keeps triggering wakeups
The "logging" AoC (Always-on Coprocessor) service is wakeup-capable, and it
triggers wakeups very frequently, thereby disrupting deep sleep on the AP:
[10839.619254] PM: pm_system_irq_wakeup: 175 triggered 18300000.mbox
...
[10839.666239] aoc 19000000.aoc: Service "logging" has 26 messages to read on wake

This hurts screen-off battery life significantly, which is unacceptable for
a logging mechanism.

Therefore, block the "logging" service to stop it from triggering frequent
wakeups. This improves screen-off battery life quite a bit.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

# Conflicts:
#	google-modules/aoc/aoc.c