pcie: exynos: Never enable clocks when reading/writing registers
clk_prepare() may sleep by acquiring a mutex lock, which means that this
code was never correct because it would try to enable clocks underneath a
spin lock.

Since there's no world where this is the correct thing to do, remove it. If
a register read/write fails because the PCIe link is down, then just let it
fail.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: dawfukfr <dawfukfr@gmail.com>