dts: gs101 : cpu : Update topology so that it is flat
Since all CPUs share the last level cache - LLC - (the DSU), the
scheduler topology should reflect that which has impact on load
balancing decisions where we've seen some tasks getting stuck on mid
cores while an idle little core is available. The load balancer
algorithm treats this scenario differently because it assumes the two
have different LLC. But in reality the point of the DSU is to make all
CPUs have the same LLC and reduce the cost of migration etc.

Flatten the topology to match the upstream recommendation.

Signed-off-by: Aarqw12 <lcockx@protonmail.com>