i2c: exynos5: Hack transfer timeout back to 100 ms after boot completes
It is observed that waking the device up from deep sleep sometimes causes
the screen to take a long time to turn on, due to a doomed i2c transfer
waiting 5 seconds for the transfer to complete.

This is awful, but abuse the cancel_boot_freq node in devfreq to know when
sys.boot_completed=1 occurs so that the transfer timeout can be lowered
back down to 100 ms. The longer timeout is only needed to ensure all i2c
peripherals successfully initialize at boot, so it's not needed once
Android confirms that everything's working.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>