eh: Use interrupts to wait for decompression completion on RT
Since the decompression path can sleep on RT, wait for decompression
completion by sleeping rather than busy waiting on RT. EH supports an IRQ
for each decompression engine to signal decompression completion, so take
advantage of it to wait exactly as long as needed until a decompression
request finishes.

On !RT, busy polling is retained as it is required.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>