gti: fix the race between irq and pm
In the context of the TDDI project, the vendor driver
will handle wake-up gestures without requiring a prior
request for an IRQ wakelock. This situation may potentially
lead to a race condition of input_process() between the
release_all_fingers() and the irq_thread_fn(). The changes
is to keep IRQ wakelock logic to avoid the race between.

Bug: 298137746
Test: verify wake-up gesture.
Test: keeps persistent fingers to reproduce issue along
      with suspend and resume.
Change-Id: I358f386fe6a428b07158c09f447dcab030ea9127
Signed-off-by: Super Liu <supercjliu@google.com>