panel: allow to update panel idle mode later
Updating the panel idle mode in the composer thread may be blocked by
the display driver when a frame is under transfer, resulting in frame
drop. To avoid this, allow the panel idle mode to be updated at the
pre-commit of the next frame, which ensures that command can be sent
without delay.

Bug: 263362191
Test: AoD, check refresh rate indicator
Test: update panel_idle, min_vrefresh and idle_delay_ms randomly
Change-Id: Ifcd2679c823c0e6aae3a8661d9f8bd209beada45
Signed-off-by: Midas Chien <midaschieh@google.com>