af_unix: Block libperfmgr from writing to the logd socket
The logcat spam from libperfmgr when the kernel doesn't have everything it
wants destroys logcats and creates excessive logd overhead from the sheer
number of logs it prints.

Block libperfmgr from writing to the logd socket so that it cannot write to
logcat at all. This way, all libperfmgr attempts to write to logcat are
killed early on before they can incur overhead, which solves the excessive
logd overhead problem.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Aarqw12 <lcockx@protonmail.com>